pyproject.toml
[project]
name = "code-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ast-grep-py>=0.26.3",
    "astra-assistants>=2.1.0.19",
    "python-fasthtml==0.4.5",
]

[tool.uv.sources]
python-fasthtml = { git = "https://github.com/AnswerDotAI/fasthtml/", rev = "00f7256453ab63a172853b0b3c604be79f01b13f" }
